refactor(why): extract PR reference parsing
